+++ Intercom.xcframework/ios-arm64/Intercom.framework/Headers/ICMHelpCenterArticleAuthor.h
@@ -10,6 +10,7 @@
 
 NS_ASSUME_NONNULL_BEGIN
 
+NS_SWIFT_SENDABLE
 NS_SWIFT_NAME(HelpCenterArticleAuthor)
 @interface ICMHelpCenterArticleAuthor : NSObject