+++ Intercom.xcframework/ios-arm64_x86_64-simulator/Intercom.framework/Headers/ICMHelpCenterArticle.h
@@ -10,6 +10,7 @@
 
 NS_ASSUME_NONNULL_BEGIN
 
+NS_SWIFT_SENDABLE
 NS_SWIFT_NAME(HelpCenterArticle)
 @interface ICMHelpCenterArticle : NSObject