+++ Intercom.xcframework/ios-arm64_x86_64-simulator/Intercom.framework/Headers/ICMHelpCenterArticleSearchResult.h
@@ -13,6 +13,7 @@ NS_ASSUME_NONNULL_BEGIN
 /**
  ICMHelpCenterArticleSearchResult represents the result of a Help Center search.
 */
+NS_SWIFT_SENDABLE
 NS_SWIFT_NAME(HelpCenterArticleSearchResult)
 @interface ICMHelpCenterArticleSearchResult : NSObject