+++ Intercom.xcframework/ios-arm64_x86_64-simulator/Intercom.framework/Headers/ICMHelpCenterCollectionContent.h
@@ -8,7 +8,6 @@
 
 #import <Foundation/Foundation.h>
 @class ICMHelpCenterArticle;
-@class ICMHelpCenterSection;
 @class ICMHelpCenterCollection;
 @class ICMHelpCenterArticleAuthor;
 
@@ -17,6 +16,7 @@ NS_ASSUME_NONNULL_BEGIN
 /**
  ICMHelpCenterCollection represents a Help Center collection and its contents. 
 */
+NS_SWIFT_SENDABLE
 NS_SWIFT_NAME(HelpCenterCollectionContent)
 @interface ICMHelpCenterCollectionContent : NSObject
 
@@ -46,12 +46,6 @@ NS_SWIFT_NAME(HelpCenterCollectionContent)
 @property (nonatomic, assign) NSInteger articleCount;
 
 /**
- @deprecated
- The sections contained in this collection.
- */
-@property (nonatomic, strong) NSArray<ICMHelpCenterSection *> *sections DEPRECATED_MSG_ATTRIBUTE("'sections' property is deprecated and will be removed in a future release. Use 'collections' instead.");
-
-/**
  The collections contained inside this collection.
  */
 @property (nonatomic, strong) NSArray<ICMHelpCenterCollection *> *collections;
@@ -60,17 +54,6 @@ NS_SWIFT_NAME(HelpCenterCollectionContent)
  The authors of articles in this collection.
  */
 @property (nonatomic, copy) NSArray<ICMHelpCenterArticleAuthor *> *authors;
-
-/**
- @deprecated
- */
-- (instancetype)initWithCollectionId:(NSString *)collectionId
-                               title:(NSString *)title
-                             summary:(nullable NSString *)summary
-                            articles:(NSArray<ICMHelpCenterArticle *> *)articles
-                   articleCount:(NSInteger)articleCount
-                            sections:(NSArray<ICMHelpCenterSection *> *)sections
-                             authors:(NSArray<ICMHelpCenterArticleAuthor *> *)authors DEPRECATED_MSG_ATTRIBUTE("'initWithCollectionId:title:summary:articles:articleCount:sections:authors:' is deprecated and will be removed in a future release. Use 'initWithCollectionId:title:summary:articles:articleCount:collections:authors:' instead.");
 
 - (instancetype)initWithCollectionId:(NSString *)collectionId
                                title:(NSString *)title